+++ CHANGELOG.md
@@ -16,6 +16,13 @@ should change the heading of the (upcoming) version to include a major version b
 
 -->
 
+# 6.1.3
+
+## @rjsf/utils
+
+- Fixed issue with default value not being prefilled when object with if/then is nested inside another object, fixing [#4222](https://github.com/rjsf-team/react-jsonschema-form/issues/4222)
+
+
 # 6.1.2
 
 ## @rjsf/antd